src/backend/requirements.txt
fastapi==0.115.14
Flask==3.1.1
Flask-Limiter==3.12.0
mysql-connector-python>=9.1.0
bcrypt==4.3.0
waitress==3.0.2
python-dotenv==1.1.1
aiomysql==0.2.0
uvicorn==0.34.3
websockets==15.0.1